LeadAgent R6 Class

A LeadAgent is an agent that can delegate tasks to specialized sub-agents. It automatically has a delegate_to_agent tool that allows it to spawn sub-agents based on registered AgentDefinitions.

Super class

deputy::Agent -> LeadAgent

Public fields

sub_agent_defs

List of AgentDefinition objects

Methods

Public methods

LeadAgent$new()

LeadAgent$register_sub_agent()

LeadAgent$available_sub_agents()

LeadAgent$print()

LeadAgent$clone()

Inherited methods

deputy::Agent$add_hook()

deputy::Agent$compact()

deputy::Agent$cost()

deputy::Agent$last_turn()

deputy::Agent$load_mcp()

deputy::Agent$load_session()

deputy::Agent$load_skill()

deputy::Agent$mcp_tools()

deputy::Agent$provider()

deputy::Agent$register_tool()

deputy::Agent$register_tools()

deputy::Agent$run()

deputy::Agent$run_shiny()

deputy::Agent$run_sync()

deputy::Agent$save_session()

deputy::Agent$settings()

deputy::Agent$skills()

deputy::Agent$slash_commands()

deputy::Agent$turns()

Method new()

Create a new LeadAgent.

Usage

LeadAgent$new(
  chat,
  sub_agents = list(),
  tools = list(),
  system_prompt = NULL,
  permissions = NULL,
  working_dir = getwd()
)

Arguments

chat

An ellmer Chat object

sub_agents

List of agent_definition() objects

tools

Additional tools for the lead agent

system_prompt

System prompt for the lead agent

permissions

Permissions for the lead agent (also applied to sub-agents)

working_dir

Working directory

Returns

A new LeadAgent object

Method register_sub_agent()

Register a new sub-agent definition.

Usage

LeadAgent$register_sub_agent(definition)

Arguments

definition

An agent_definition() object

Returns

Invisible self

Method available_sub_agents()

Get available sub-agent names.

Usage

LeadAgent$available_sub_agents()

Returns

Character vector of sub-agent names

Method print()

Print the lead agent.

Usage

LeadAgent$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

LeadAgent$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.